move build to its own function
