+++ packages/cli/src/runDeploy.ts
@@ -7,27 +7,14 @@ import { privateKeyToAccount } from "viem/accounts";
 import { loadConfig } from "@latticexyz/config/node";
 import { StoreConfig } from "@latticexyz/store";
 import { WorldConfig } from "@latticexyz/world";
-import {
-  forge,
-  getForgeConfig,
-  getOutDirectory,
-  getRemappings,
-  getRpcUrl,
-  getSrcDirectory,
-} from "@latticexyz/common/foundry";
+import { getOutDirectory, getRpcUrl, getSrcDirectory } from "@latticexyz/common/foundry";
 import chalk from "chalk";
-import { execa } from "execa";
 import { MUDError } from "@latticexyz/common/errors";
 import { resolveConfig } from "./deploy/resolveConfig";
 import { getChainId } from "viem/actions";
 import { postDeploy } from "./utils/utils/postDeploy";
 import { WorldDeploy } from "./deploy/common";
-import { tablegen } from "@latticexyz/store/codegen";
-import { worldgen } from "@latticexyz/world/node";
-import { getExistingContracts } from "./utils/getExistingContracts";
-import { debug as parentDebug } from "./debug";
-
-const debug = parentDebug.extend("runDeploy");
+import { build } from "./build";
 
 export const deployOptions = {
   configPath: { type: "string", desc: "Path to the config file" },
@@ -60,7 +47,6 @@ export async function runDeploy(opts: DeployOptions): Promise<WorldDeploy> {
 
   const srcDir = opts.srcDir ?? (await getSrcDirectory(profile));
   const outDir = await getOutDirectory(profile);
-  const remappings = await getRemappings();
 
   const rpc = opts.rpc ?? (await getRpcUrl(profile));
   console.log(
@@ -71,24 +57,7 @@ export async function runDeploy(opts: DeployOptions): Promise<WorldDeploy> {
 
   // Run build
   if (!opts.skipBuild) {
-    const outPath = path.join(srcDir, config.codegenDirectory);
-    await Promise.all([tablegen(config, outPath, remappings), worldgen(config, getExistingContracts(srcDir), outPath)]);
-
-    // TODO remove when https://github.com/foundry-rs/foundry/issues/6241 is resolved
-    const forgeConfig = await getForgeConfig(profile);
-    if (forgeConfig.cache) {
-      const cacheFilePath = path.join(forgeConfig.cache_path, "solidity-files-cache.json");
-      if (existsSync(cacheFilePath)) {
-        debug("Unsetting cached content hash of IWorld.sol to force it to regenerate");
-        const solidityFilesCache = JSON.parse(readFileSync(cacheFilePath, "utf8"));
-        const worldInterfacePath = path.join(outPath, "world", "IWorld.sol");
-        solidityFilesCache["files"][worldInterfacePath]["contentHash"] = "";
-        writeFileSync(cacheFilePath, JSON.stringify(solidityFilesCache, null, 2));
-      }
-    }
-
-    await forge(["build"], { profile });
-    await execa("mud", ["abi-ts"], { stdio: "inherit" });
+    await build({ config, srcDir, foundryProfile: profile });
   }
 
   const privateKey = process.env.PRIVATE_KEY as Hex;
